1.Device和Data_history使用MongoRepository进行数据管理 2.接口及其相关内容修改

+++ src/main/java/com/hydata/intelligence/platform/controller/DataStreamModelController.java
@@ -88,7 +88,7 @@ public class DataStreamModelController {
 	 }
 	 
 	 @RequestMapping(value ="/get_increment",method = RequestMethod.GET)
-	 public JSONObject getIncrement(Integer product_id,String start,String end) throws ParseException {		 
+	 public JSONObject getIncrement(Long product_id,String start,String end) throws ParseException {		 
 	     return dsmService.getIncrement(product_id, sdf.parse(start), sdf.parse(end));
 	 }
